# perf(turbopack): Migrate as many data structures to FxHash as possible