--- a/crates/next-core/src/next_edge/route_regex.rs
+++ b/crates/next-core/src/next_edge/route_regex.rs
@@ -1,10 +1,9 @@
 //! The following code was mostly generated using GTP-4 from
 //! next.js/packages/next/src/shared/lib/router/utils/route-regex.ts
 
-use std::collections::HashMap;
-
 use once_cell::sync::Lazy;
 use regex::Regex;
+use rustc_hash::FxHashMap;
 
 const INTERCEPTION_ROUTE_MARKERS: [&str; 4] = ["(..)(..)", "(.)", "(..)", "(...)"];
 const NEXT_QUERY_PARAM_PREFIX: &str = "nxtP";
@@ -19,7 +18,7 @@ pub struct Group {
 
 #[derive(Debug)]
 pub struct RouteRegex {
-    pub groups: HashMap<String, Group>,
+    pub groups: FxHashMap<String, Group>,
     pub regex: String,
 }
 
@@ -27,7 +26,7 @@ pub struct RouteRegex {
 pub struct NamedRouteRegex {
     pub regex: RouteRegex,
     pub named_regex: String,
-    pub route_keys: HashMap<String, String>,
+    pub route_keys: FxHashMap<String, String>,
 }
 
 #[derive(Debug)]
@@ -83,9 +82,9 @@ fn remove_trailing_slash(route: &str) -> &str {
 
 static PARAM_MATCH_REGEX: Lazy<Regex> = Lazy::new(|| Regex::new(r"\[((?:\[.*\])|.+)\]").unwrap());
 
-fn get_parametrized_route(route: &str) -> (String, HashMap<String, Group>) {
+fn get_parametrized_route(route: &str) -> (String, FxHashMap<String, Group>) {
     let segments: Vec<&str> = remove_trailing_slash(route)[1..].split('/').collect();
-    let mut groups: HashMap<String, Group> = HashMap::new();
+    let mut groups: FxHashMap<String, Group> = FxHashMap::default();
     let mut group_index = 1;
     let parameterized_route = segments
         .iter()
@@ -163,7 +162,7 @@ fn build_get_safe_route_key() -> impl FnMut() -> String {
 fn get_safe_key_from_segment(
     get_safe_route_key: &mut impl FnMut() -> String,
     segment: &str,
-    route_keys: &mut HashMap<String, String>,
+    route_keys: &mut FxHashMap<String, String>,
     key_prefix: Option<&'static str>,
 ) -> String {
     let ParsedParameter {
@@ -207,10 +206,10 @@ fn get_safe_key_from_segment(
 fn get_named_parametrized_route(
     route: &str,
     prefix_route_keys: bool,
-) -> (String, HashMap<String, String>) {
+) -> (String, FxHashMap<String, String>) {
     let segments: Vec<&str> = remove_trailing_slash(route)[1..].split('/').collect();
     let get_safe_route_key = &mut build_get_safe_route_key();
-    let mut route_keys: HashMap<String, String> = HashMap::new();
+    let mut route_keys: FxHashMap<String, String> = FxHashMap::default();
     let parameterized_route = segments
         .iter()
         .map(|segment| {