--- a/crates/next-custom-transforms/src/lib.rs
+++ b/crates/next-custom-transforms/src/lib.rs
@@ -32,6 +32,13 @@ DEALINGS IN THE SOFTWARE.
 #![feature(arbitrary_self_types)]
 #![feature(arbitrary_self_types_pointers)]
 
+use std::hash::BuildHasherDefault;
+
+use indexmap::IndexMap;
+use rustc_hash::FxHasher;
+
 pub mod chain_transforms;
 mod linter;
 pub mod transforms;
+
+type FxIndexMap<K, V> = IndexMap<K, V, BuildHasherDefault<FxHasher>>;