… Cargo.lock
… crates/next-api/Cargo.toml
… crates/next-custom-transforms/src/transforms/react_server_components.rs
… crates/next-error-code-swc-plugin/Cargo.lock
… crates/next-error-code-swc-plugin/Cargo.toml
… crates/next-error-code-swc-plugin/src/lib.rs
… turbopack/crates/turbo-static/Cargo.toml
… turbopack/crates/turbo-static/src/call_resolver.rs
… turbopack/crates/turbo-static/src/main.rs
… turbopack/crates/turbo-tasks-backend/src/backend/mod.rs
… turbopack/crates/turbo-tasks-build/Cargo.toml
… turbopack/crates/turbo-tasks-build/src/lib.rs
… turbopack/crates/turbo-tasks-fs/Cargo.toml
… turbopack/crates/turbo-tasks-fs/src/invalidator_map.rs
… turbopack/crates/turbo-tasks-fs/src/lib.rs
… turbopack/crates/turbo-tasks-fs/src/mutex_map.rs
… turbopack/crates/turbo-tasks-fs/src/read_glob.rs
… turbopack/crates/turbo-tasks-fs/src/watcher.rs
… turbopack/crates/turbo-tasks-macros/Cargo.toml
… turbopack/crates/turbo-tasks-macros/src/func.rs
… turbopack/crates/turbo-tasks-macros/src/generic_type_macro.rs
… turbopack/crates/turbo-tasks-memory/src/aggregation/tests.rs
… turbopack/crates/turbo-tasks-testing/src/lib.rs
… turbopack/crates/turbo-tasks-testing/tests/collectibles.rs
… turbopack/crates/turbo-tasks/src/debug/mod.rs
… turbopack/crates/turbo-tasks/src/effect.rs
… turbopack/crates/turbo-tasks/src/graph/adjacency_map.rs
… turbopack/crates/turbo-tasks/src/graph/graph_store.rs
… turbopack/crates/turbo-tasks/src/graph/graph_traversal.rs
… turbopack/crates/turbopack-bench/Cargo.toml