Migrate as many things as possible that I could find with:

```
rg '\b(Index|Hash)(Map|Set)<'
```

# Benchmarks

I tried building next-site, differences are likely within noise threshold.

## Before this change

```
$ hyperfine --warmup 1 --runs 5 --prepare 'rm -rf .next' 'TURBOPACK=1 TURBOPACK_BUILD=1 TURBO_ENGINE_READ_ONLY=1 pnpm next build --experimental-build-mode=compile'

Benchmark 1: TURBOPACK=1 TURBOPACK_BUILD=1 TURBO_ENGINE_READ_ONLY=1 pnpm next build --experimental-build-mode=compile
  Time (mean ± σ):     27.254 s ±  0.268 s    [User: 125.055 s, System: 16.090 s]
  Range (min … max):   26.810 s … 27.529 s    5 runs
```

## After this change

```
$ hyperfine --warmup 1 --runs 5 --prepare 'rm -rf .next' 'TURBOPACK=1 TURBOPACK_BUILD=1 TURBO_ENGINE_READ_ONLY=1 pnpm next build --experimental-build-mode=compile'

Benchmark 1: TURBOPACK=1 TURBOPACK_BUILD=1 TURBO_ENGINE_READ_ONLY=1 pnpm next build --experimental-build-mode=compile
  Time (mean ± σ):     26.711 s ±  0.461 s    [User: 123.910 s, System: 15.667 s]
  Range (min … max):   26.346 s … 27.463 s    5 runs
```